[feature] creating event call

+++ app/src/main/java/com/example/fusmobilni/model/event/EventType.java
@@ -7,20 +7,20 @@ import com.example.fusmobilni.model.items.category.OfferingsCategory;
 import java.util.List;
 
 public class EventType {
-    private final Integer id;
+    private final Long id;
     private String name;
     private String description;
     private List<OfferingsCategory> suggestedCategories;
 
     private boolean isActive=true;
-    public EventType(Integer id,String name, String description, List<OfferingsCategory> categories) {
+    public EventType(Long id,String name, String description, List<OfferingsCategory> categories) {
         this.id = id;
         this.name = name;
         this.description = description;
         suggestedCategories = categories;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
